requirements.txt
wheel
ipdb
ipykernel
yapf
matplotlib
pytest
seaborn
numpy
pandas
scipy
scikit-learn
google.cloud
google-cloud-bigquery
google-cloud-storage
requests
db-dtypes
pyarrow
colorama
typing
# Trick to install the version of Tensorflow depending on your processor: darwin == Mac, ARM == M1
# tensorflow-macos==2.10.0; sys_platform == 'darwin' and 'ARM' in platform_version # Mac M-chips
# tensorflow==2.10.0; sys_platform == 'darwin' and 'ARM' not in platform_version # Mac Intel chips
# tensorflow==2.10.0; sys_platform != 'darwin' # Windows & Ubuntu Intel chips

# prevent bq 3+ db-dtypes package error when running training from bq

mlflow
prefect

python-dotenv
psycopg2-binary

# API
fastapi
pytz
uvicorn

# ML
xgboost

# Class programming
abc
